LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/position-area-wm-dir-expected.txt


FAIL Offsets for: start start with writing-mode / direction: containing-block: horizontal-tb / rtl assert_equals: Checking offsetLeft expected 250 but got 400
FAIL Offsets for: self-start self-start with writing-mode / direction: containing-block: horizontal-tb / rtl assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: x-start y-start with writing-mode / direction: containing-block: horizontal-tb / rtl assert_equals: Checking offsetLeft expected 250 but got 400
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: containing-block: horizontal-tb / rtl assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: inline-start block-end with writing-mode / direction: containing-block: horizontal-tb / rtl assert_equals: Checking offsetLeft expected 250 but got 400
FAIL Offsets for: self-inline-start self-block-end with writing-mode / direction: containing-block: horizontal-tb / rtl assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: start start with writing-mode / direction: containing-block: vertical-lr / ltr assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: self-start self-start with writing-mode / direction: containing-block: vertical-lr / ltr assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: x-start y-start with writing-mode / direction: containing-block: vertical-lr / ltr assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: containing-block: vertical-lr / ltr assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: inline-start block-end with writing-mode / direction: containing-block: vertical-lr / ltr assert_equals: Checking offsetLeft expected 250 but got 400
FAIL Offsets for: self-inline-start self-block-end with writing-mode / direction: containing-block: vertical-lr / ltr assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: start start with writing-mode / direction: containing-block: vertical-lr / rtl assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: self-start self-start with writing-mode / direction: containing-block: vertical-lr / rtl assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: x-start y-start with writing-mode / direction: containing-block: vertical-lr / rtl assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: containing-block: vertical-lr / rtl assert_equals: Checking offsetLeft expected 0 but got 400
FAIL Offsets for: start start with writing-mode / direction: containing-block: vertical-rl / ltr assert_equals: Checking offsetLeft expected 250 but got 0
FAIL Offsets for: self-start self-start with writing-mode / direction: containing-block: vertical-rl / ltr assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: x-start y-start with writing-mode / direction: containing-block: vertical-rl / ltr assert_equals: Checking offsetLeft expected 250 but got 0
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: containing-block: vertical-rl / ltr assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: start start with writing-mode / direction: containing-block: vertical-rl / rtl assert_equals: Checking offsetLeft expected 250 but got 0
FAIL Offsets for: self-start self-start with writing-mode / direction: containing-block: vertical-rl / rtl assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: x-start y-start with writing-mode / direction: containing-block: vertical-rl / rtl assert_equals: Checking offsetLeft expected 250 but got 0
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: containing-block: vertical-rl / rtl assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: start start with writing-mode / direction: self: horizontal-tb / rtl assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: self-start self-start with writing-mode / direction: self: horizontal-tb / rtl assert_equals: Checking offsetLeft expected 250 but got 0
FAIL Offsets for: x-start y-start with writing-mode / direction: self: horizontal-tb / rtl assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: self: horizontal-tb / rtl assert_equals: Checking offsetLeft expected 250 but got 0
FAIL Offsets for: start start with writing-mode / direction: self: vertical-lr / ltr assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: self-start self-start with writing-mode / direction: self: vertical-lr / ltr assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: x-start y-start with writing-mode / direction: self: vertical-lr / ltr assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: self: vertical-lr / ltr assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: start start with writing-mode / direction: self: vertical-lr / rtl assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: self-start self-start with writing-mode / direction: self: vertical-lr / rtl assert_equals: Checking offsetTop expected 225 but got 0
FAIL Offsets for: x-start y-start with writing-mode / direction: self: vertical-lr / rtl assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: self: vertical-lr / rtl assert_equals: Checking offsetTop expected 225 but got 0
FAIL Offsets for: start start with writing-mode / direction: self: vertical-rl / ltr assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: self-start self-start with writing-mode / direction: self: vertical-rl / ltr assert_equals: Checking offsetLeft expected 250 but got 0
FAIL Offsets for: x-start y-start with writing-mode / direction: self: vertical-rl / ltr assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: self: vertical-rl / ltr assert_equals: Checking offsetLeft expected 250 but got 0
FAIL Offsets for: start start with writing-mode / direction: self: vertical-rl / rtl assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: self-start self-start with writing-mode / direction: self: vertical-rl / rtl assert_equals: Checking offsetLeft expected 250 but got 0
FAIL Offsets for: x-start y-start with writing-mode / direction: self: vertical-rl / rtl assert_equals: Checking offsetWidth expected 100 but got 0
FAIL Offsets for: x-self-start y-self-start with writing-mode / direction: self: vertical-rl / rtl assert_equals: Checking offsetLeft expected 250 but got 0
